Test for weird behaviour on the wins-work pr
This pr opens up a url on the live site under hackforla.org/wins-test/
Which exposes the winsTest directory which is a duplicate of the wins directory.

This is being done to try and replicate a weird behavoiour that was seen previously on the live page
when the pr 'wins-work' was merged.