Fixed bug where dynamic text was not outside <span>
While the code does not look pretty, it is unfortunately the best way I could find to keep the dynamic text card[team] and card[role] outside of the span elements, without using innerHTML.

The reason we are avoiding using innerHTML is to protect against XSS vulnerabilities. However, it makes it complicated to add <span> text within a <div>, and then additional text outside of the <span> but still within the <div>.

When using appendChild(), the method returns the new child, rather than the original querySelector() element. Thus if you are chaining appendChild() after each other, the second appendChild() will be placed within the first.

Code is separated into separate sections:
(1) Creating the elements
(2) Preparing the text of the elements
(3) Inserting the elements into the DOM