refactor: add github-handle for Hyun Joo Sandy Oh
We need to create a single variable github-handle to hold the github
handle for each member of the leadership team. Eventually github-handle
will replace the github and picture variables, reducing redundancy in
the project file.

In _projects/youthjusticenav.md, this is done by adding a key,
github-handle, to the leadership item, Hyun Joo Sandy Oh.