Creates new Flag Issues Unlabeled After Deletion workflow, with JS hooks
This new workflow is called by Update Label Directory if a label is deleted, and creates a notification issue if the deleted label was applied to any other issues in the repo. It then posts a comment to the agenda issue linking to the notification and providing some high level details.

If the agenda issue isn't present, the workflow will create a new issue notifying that the agenda issue is missing.